Updating simulation and para simulation functions to accomodate the parameter configuration functionality. Still work in progress:
will add an if statement where both data and parameter simulations take place. Adjust both simulate and para_simulate.


To account for binary gamma parameter:
Make the configuration table with the first or last column is the gamma = 1 or 2. Then repeat the other configuration combos with 1 and 2 in the column. Then for the simulate function it is easier for the function to index through columns and not need a separate section for it. OOOOr add a argin that specifies 1 or 2. 

Remember to update all scripts that use para or simulate to ensure correct number of inputs and outputs.